Random number generator fix
Tried to hold the maximum value for a pointer in just a normal unsigned long. It was a worthy attempt, but the variable just couldn't handle it. Changed it to the (supposedly) correct type of intptr_t.

+++ src/compat.cpp
@@ -153,7 +153,7 @@ void pause_ms(int t)
 }
 
 // FNV-1a 32-bit hash function (fast and effective) -- helper function for getSeed()
-void fnvHash(unsigned long &fnv_hash,unsigned long num)
+void fnvHash(unsigned long &fnv_hash,intptr_t num)
 {
    fnv_hash*=16777619UL; // multiply by the FNV-1a hash's 32-bit prime
    fnv_hash&=0xffffffffUL; // keep the number 32-bit (we could be on a 64-bit system)
